add plugin event for processing messages posted by the bot
Instead of throwing out messages sent by the bot, forward them to plugins
using a new event (channel_message_delivered). The reasoning is that most
plugins don't care about outgoing messages and duplicating the "is it my
post" logic in each plugin would be error-prone. On the other hand,
channel_message_edited is called for both the bot's and other users'
messages, as it is assumed that plugins that care about edited messages are
more likely to be e.g. logging plugins and therefore care about situations
where both the bot and the user edit their messages.