grammargen: allow for result memoization
To make some consistency possible in the results of grammars, add an "@"
flag to rules which marks the rule as memoized. Then, when this rule is,
called, the result is remembered at the call site and returned every
subsequent time the same call site is reached. To simulate memoization
at the rule itself (instead of at the call site), add a layer of
indirection, memoizing the inner rule but calling only the outer one.