bim: protect against multiple instances of the same vehicle
The user may try to list a vehicle multiple times, or a vehicle might be
counted both due to explicit specification and due to membership in a
fixed coupling. Since it is impossible to ride a vehicle multiple times
as part of one ride, the primary key in the database has been defined
accordingly. To ensure the registration of the ride doesn't fail with a
constraint violation error, ensure uniqueness of each vehicle entry
within the ride while still correctly differentiating between explicitly
and implicitly registered vehicles.