bim: don't overcount lone/last bims, attempt #2
Fix the !lonebims and !lastbims queries to only count distinct
vehicles, i.e. not count the same vehicle multiple times even if the
ridden it multiple times.

This version ensures that the query is compatible with PostgreSQL, which
apparently doesn't accept multiple columns after the DISTINCT specifier
in the arguments to an aggregate function.