sed: reset index when parsing new replacement grp
The "\g12;" syntax allows replacing groups beyond 9. Thanks to an
oversight, the number of the group wasn't reset between each occurrence
of this syntax, so a "\g12;" followed by a "\g13;" later in the string
would be interpreted as references to groups 12 and 1213, respectively.
Fix this to refer to groups 12 and 13, as obviously intended.